#6977 Garantie bei Produkten https://github.com/metasfresh/metasfresh/issues/10648
