+++ backend/de.metas.business/src/main/java/de/metas/product/IProductDAO.java
@@ -173,4 +173,5 @@ public interface IProductDAO extends ISingletonService
 
 	int getProductGuaranteeDaysMinFallbackProductCategory(@NonNull final ProductId productId);
 
+	int getGuaranteeMonthsInDays(I_M_Product product);
 }
